requirements.txt
google-generativeai==0.5.2
python-dotenv==1.0.1
requests==2.32.3
qrcode==7.4.2
Pillow==10.4.0
